+++ components/blog/post.module.scss
@@ -6,7 +6,7 @@
 	gap: $space-xs;
 	background-color: rgb(255, 255, 255);
 	border: 1px solid rgb(228, 228, 228);
-	padding: $space-md;
+	padding: $space-sm $space-md;
 	border-radius: $space-sm;
 	font-family: $primaryTextFont;
 	h3 {
