+++ pages/blog/edit/[id].tsx
@@ -17,7 +17,7 @@ function EditPost(props: any) {
 			</section>
 			<section className={styles.rightCol}>
 				<div className={styles.back}>
-					<Link href="/" passHref>
+					<Link href="/dashboard" passHref>
 						<button className={styles.backButton}>
 							<FaChevronLeft /> Back{' '}
 						</button>
