+++ pages/blog/edit/edit.module.scss
@@ -5,6 +5,8 @@
 	grid-template-columns: 300px 1fr 180px;
 	margin: $space-sm 0;
 	justify-items: center;
+	max-width: 1200px;
+	margin: $space-lg auto 0 auto;
 
 	.leftCol {
 		display: flex;
