+++ pages/blog/index.module.scss
@@ -5,6 +5,8 @@
 	grid-template-columns: 1fr 180px;
 	gap: 20px;
 	margin-top: $space-lg;
+	max-width: 1200px;
+	margin: $space-lg auto 0 auto;
 	:first-child{
 		max-width: 700px;
 		justify-self: center;
