+++ pages/dashboard.module.scss
@@ -1,11 +1,12 @@
-@import '../../styles/globals.scss';
+@import '../styles/globals.scss';
 
 .container {
 	display: grid;
 	grid-template-columns: 1fr 180px;
 	gap: 20px;
 	margin-top: $space-lg;
-
+	max-width: 1200px;
+	margin: $space-lg auto 0 auto;
 }
 
 .main {
