+++ pages/index.tsx
@@ -1,48 +1,29 @@
 import React from 'react';
-import { useSession } from 'next-auth/react';
-import { PrismaClient } from '@prisma/client';
+import { getSession } from 'next-auth/react';
 import { Login } from '../components/auth/login';
 import styles from './index.module.scss';
-import Dashboard from '../components/dashboard';
 
-type Props = {
-	posts: any;
-}
-
-function Home(props: Props) {
-	const { posts } = props;
-	const { data: session } = useSession();
+function Home() {
 	return (
-		(session
-			? <Dashboard posts={posts} />
-			:			(
-				<div className={styles.login}>
-					<Login />
-				</div>
-			))
-
+		<div className={styles.login}>
+			<Login />
+		</div>
 	);
 }
-
-export async function getServerSideProps() {
-	const prisma = new PrismaClient();
-	// Fetch all posted jobs and include related items from Company table
-	const posts = await prisma.blog.findMany({
-		take: 3,
-		orderBy: [
-			{
-				date: 'desc',
+export async function getServerSideProps(context: { req: any; }) {
+	const session = await getSession({ req: context.req });
+	// Redirect if user isn't logged in
+	if (session) {
+		return {
+			redirect: {
+				destination: '/dashboard',
+				permanent: false,
 			},
-		],
-	});
-	console.log(posts);
+		};
+	}
+
 	return {
-		props: {
-			posts: posts.map((post) => ({
-				...post,
-				date: post.date.toISOString(),
-			})),
-		},
+		props: { session },
 	};
 }
