+++ pages/spotlight/create.tsx
@@ -13,7 +13,7 @@ function CreateSpotlight() {
 			</section>
 			<section className={styles.rightCol}>
 				<div className={styles.back}>
-					<Link href="/" passHref>
+					<Link href="/dashboard" passHref>
 						<button className={styles.backButton}><FaChevronLeft /> Back </button>
 					</Link>
 				</div>
